requirements.txt
Django==2.0.6
djangorestframework==3.8.2
factory-boy==2.11.1
psycopg2==2.7.5
